Rearranged tests
